refactor application service's orphan data function
This function is now getting very long and is almost a service in it's own right.
