+++ server/services/applicationService.ts
@@ -7,7 +7,7 @@ import { getApplicationSubmissionData, getApplicationUpdateData } from '../utils
 import TaskListPage, { TaskListPageInterface } from '../form-pages/taskListPage'
 import CheckYourAnswers from '../form-pages/apply/check-your-answers/check-your-answers/checkYourAnswers'
 import { ValidationError } from '../utils/errors'
-import { lastKnownKeys, previousKeys } from '../form-pages/apply/about-the-person/address-history/previousAddress'
+import deleteOrphanedFollowOnAnswers from '../utils/applications/deleteOrphanedData'
 
 export default class ApplicationService {
   constructor(private readonly applicationClientFactory: RestClientBuilder<ApplicationClient>) {}
@@ -63,7 +63,7 @@ export default class ApplicationService {
       const oldBody = application.data?.[taskName]?.[pageName]
 
       application.data = this.addPageDataToApplicationData(application.data, taskName, pageName, page)
-      application.data = this.deleteOrphanedFollowOnAnswers(application.data)
+      application.data = deleteOrphanedFollowOnAnswers(application.data)
       application.data = this.deleteCheckYourAnswersIfPageChange(application.data, pageName, oldBody, page.body)
 
       await client.update(application.id, getApplicationUpdateData(application))
@@ -99,86 +99,6 @@ export default class ApplicationService {
       if (pageHasChangedSinceLastSave()) {
         delete applicationData[checkYourAnswersTaskName]
       }
-    }
-
-    return applicationData
-  }
-
-  private deleteOrphanedFollowOnAnswers(applicationData: AnyValue): AnyValue {
-    const deleteOrphanedFundingInformation = () => {
-      delete applicationData['funding-information'].identification
-      delete applicationData['funding-information']['alternative-identification']
-    }
-
-    const deleteOrphanedEqualityInformation = () => {
-      Object.keys(applicationData['equality-and-diversity-monitoring']).forEach(key => {
-        if (key !== 'will-answer-equality-questions') {
-          delete applicationData['equality-and-diversity-monitoring'][key]
-        }
-      })
-    }
-
-    const deleteOrphanedOffendingHistoryInformation = () => {
-      delete applicationData['offending-history']['offence-history-data']
-    }
-
-    const deleteAddressHistoryInformation = () => {
-      if (applicationData['address-history']['previous-address'].hasPreviousAddress === 'yes') {
-        lastKnownKeys.forEach(key => delete applicationData['address-history']['previous-address'][key])
-      } else if (applicationData['address-history']['previous-address'].hasPreviousAddress === 'no') {
-        previousKeys.forEach(key => delete applicationData['address-history']['previous-address'][key])
-      }
-    }
-
-    const hasOrphanedInformation = ({
-      taskName,
-      pageName,
-      questionKey,
-      answerToCheck,
-    }: {
-      taskName: string
-      pageName: string
-      questionKey: string
-      answerToCheck: string
-    }) => {
-      return applicationData[taskName]?.[pageName]?.[questionKey] === answerToCheck
-    }
-
-    if (
-      hasOrphanedInformation({
-        taskName: 'funding-information',
-        pageName: 'funding-source',
-        questionKey: 'fundingSource',
-        answerToCheck: 'personalSavings',
-      })
-    ) {
-      deleteOrphanedFundingInformation()
-    }
-
-    if (
-      hasOrphanedInformation({
-        taskName: 'equality-and-diversity-monitoring',
-        pageName: 'will-answer-equality-questions',
-        questionKey: 'willAnswer',
-        answerToCheck: 'no',
-      })
-    ) {
-      deleteOrphanedEqualityInformation()
-    }
-
-    if (
-      hasOrphanedInformation({
-        taskName: 'offending-history',
-        pageName: 'any-previous-convictions',
-        questionKey: 'hasAnyPreviousConvictions',
-        answerToCheck: 'no',
-      })
-    ) {
-      deleteOrphanedOffendingHistoryInformation()
-    }
-
-    if (applicationData['address-history']?.['previous-address']?.hasPreviousAddress) {
-      deleteAddressHistoryInformation()
     }
 
     return applicationData
